+++ PyBTSteward/decode_eddystone.py
@@ -9,14 +9,14 @@ import subprocess
 import sys
 import time
 from collections import namedtuple
-import PyBeacon.wpl_cfg_parser
-import PyBeacon.wpl_log
-import PyBeacon.wpl_stats
+import PyBTSteward.wpl_cfg_parser
+import PyBTSteward.wpl_log
+import PyBTSteward.wpl_stats
 import bluetooth._bluetooth as bluez
 import uuid
 from . import __version__
 from pprint import pprint
-from PyBeacon.wpl_cfg_parser import wpl_cfg
+from PyBTSteward.wpl_cfg_parser import wpl_cfg
 logger = logging.getLogger(__name__)
 
 def decode_eddystone(state, config, ad_struct):
@@ -119,7 +119,7 @@ def decode_eddystone(state, config, ad_struct):
                 except TypeError:
                     logger.debug('interpolating Eddystone UID instance from string')
                     ret['instance'] = ''.join('%02X' % ord(c) for c in ei.instance)
-                ret['rssi_ref'] = ei.rssi_ref - 41
+                ret['rssi_ref'] = ei.rssi_ref
 
             # Is this a URL sub type?
             if ec.sub_type == 0x10: