+++ PyBTSteward/wpl_cfg_parser.py
@@ -8,7 +8,7 @@ Who doesn't like Config Files.
 '''
 import yaml
 from pprint import pprint
-from PyBeacon.dict_utils import smerge_dicts, merge_dict
+from PyBTSteward.dict_utils import smerge_dicts, merge_dict
 import logging
 logger = logging.getLogger(__name__)