+++ PyBTSteward/wpl_log.py
@@ -11,7 +11,7 @@ Who doesn't like log.
 '''
 import logging
 import time
-from PyBeacon.wpl_cfg_parser import wpl_cfg
+from PyBTSteward.wpl_cfg_parser import wpl_cfg
 #https://docs.python.org/2/library/logging.html#levels
 conf = wpl_cfg()
 logging.Formatter.converter = time.localtime