+++ chef/pybtsteward/attributes/default.rb
@@ -1,0 +1,8 @@
+default['PyBTSteward']['install']['enable'] = true
+default['PyBTSteward']['install']['bluez']['enable'] = true
+default['PyBTSteward']['install']['repo']
+default['PyBTSteward']['user']
+default['PyBTSteward']['group']
+default['PyBTSteward']['basedir']
+default['PyBTSteward']['local_cfg']['manage']
+default['PyBTSteward']['local_cfg']['type'] #file/template