+++ chef/pybtsteward/metadata.rb
@@ -1,7 +1,7 @@
-name             'pybeacon'
+name             'PyBTSteward'
 maintainer       'YOUR_COMPANY_NAME'
 maintainer_email 'YOUR_EMAIL'
 license          'All rights reserved'
-description      'Installs/Configures pybeacon'
+description      'Installs/Configures PyBTSteward'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'